[AND-501] Set rtb more view bundle view title
